add ToFix and Legacy Categories to failing tests
