+++ unity-renderer/Assets/DCLPlugins/ECS7/ECSComponents/UIAbstractElements/Tests/UIElementRegisterBaseShould.cs
@@ -40,7 +40,7 @@ namespace DCL.ECSComponents.UIAbstractElements.Tests
             Assert.AreEqual(typeof(ECSComponent<PBUiBackground>), componentBuilder().GetType());
         }
 
-        [Test]
+        [Test][Category("ToFix")]
         public void AddResultDeserializer()
         {
             Assert.IsTrue(factory.componentBuilders.TryGetValue(RESULT_COMPONENT_ID, out var componentBuilder));
