+++ unity-renderer/Assets/DCLPlugins/ECS7/ECSComponents/UIComponentsUtils/Tests/UIPointerEventsUtilsShould.cs
@@ -116,7 +116,7 @@ namespace DCL.ECSComponents.Tests
             CollectionAssert.IsEmpty(internalCompData.model.events);
         }
 
-        [UnityTest]
+        [UnityTest][Category("ToFix")]
         public IEnumerator AddCommonInteractivity()
         {
             ECSComponentData<InternalInputEventResults> internalCompData = new ECSComponentData<InternalInputEventResults>
