--- unity-renderer/Assets/Scripts/Tests/EditorTests.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 2292b20a9fc1839448581ab0d1a3efd2
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
