--- unity-renderer/Assets/Scripts/Tests/EditorTests/EditorTests.asmdef
@@ -1,24 +1,0 @@
-{
-    "name": "EditorTests",
-    "rootNamespace": "",
-    "references": [
-        "GUID:6ebd21b5b1cb88347b078d9a0ce1b921",
-        "GUID:593a5b492d29ac6448b1ebf7f035ef33",
-        "GUID:9e24947de15b9834991c9d8411ea37cf",
-        "GUID:69448af7b92c7f342b298e06a37122aa",
-        "GUID:f0c774e7dbab37641808f75bcc438e4a",
-        "GUID:536993f12b814234a8f1d38fa0a1f35c",
-        "GUID:c5ecc461727906345a35491a0440694f"
-    ],
-    "includePlatforms": [
-        "Editor"
-    ],
-    "excludePlatforms": [],
-    "allowUnsafeCode": false,
-    "overrideReferences": false,
-    "precompiledReferences": [],
-    "autoReferenced": true,
-    "defineConstraints": [],
-    "versionDefines": [],
-    "noEngineReferences": false
-}
