--- unity-renderer/Assets/Scripts/Tests/EditorTests/EditorTests.asmdef.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 6ec19900385bc034290badaf1492d086
-AssemblyDefinitionImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
