+++ unity-renderer/Assets/UnityGLTF/Scripts/Tests/GLTFSceneImporterShould.cs
@@ -15,7 +15,7 @@ public class GLTFSceneImporterShould
 {
     private const string FILE_NAME = "Trunk.gltf";
     private readonly string rootDirectoryPath = Application.dataPath + "/../TestResources/GLTF/Trunk/";
-    
+
     private GLTFSceneImporter gltfLoader;
     private FileStream stream;
     private CancellationTokenSource cancellationTokenSource;
@@ -46,7 +46,7 @@ public class GLTFSceneImporterShould
         stream.Dispose();
         Object.DestroyImmediate(gltfLoader.CreatedObject);
     }
-    
+
     [UnityTest]
     public IEnumerator TrackLoadingGLTFCorrectly()
     {
@@ -58,14 +58,14 @@ public class GLTFSceneImporterShould
 
         cancellationTokenSource.Cancel();
         yield return new WaitUntil(() => task.IsCompleted);
-        
+
         (int loading, int failed, int cancelled, int loaded) = PerformanceAnalytics.GLTFTracker.GetData();
         Assert.AreEqual(0, loaded, "Loaded GLTFs");
         Assert.AreEqual(0, loading, "Loading GLTFs");
         Assert.AreEqual(0, failed, "Failed GLTFs");
         Assert.AreEqual(1, cancelled, "Cancelled GLTFs");
     }
-    
+
     [UnityTest]
     public IEnumerator TrackLoadedGLTFCorrectly()
     {
@@ -79,7 +79,7 @@ public class GLTFSceneImporterShould
         Assert.AreEqual(0, failed, "Failed GLTFs");
         Assert.AreEqual(0, cancelled, "Cancelled GLTFs");
     }
-    
+
     [UnityTest]
     public IEnumerator TrackCancelledGLTFCorrectly()
     {
@@ -90,7 +90,7 @@ public class GLTFSceneImporterShould
         Assert.AreEqual(1, PerformanceAnalytics.GLTFTracker.GetData().loading, "Loading GLTFs");
 
         cancellationTokenSource.Cancel();
-        
+
         yield return new WaitUntil(() => task.IsCompleted);
 
         (int loading, int failed, int cancelled, int loaded) = PerformanceAnalytics.GLTFTracker.GetData();
@@ -99,15 +99,15 @@ public class GLTFSceneImporterShould
         Assert.AreEqual(0, failed, "Failed GLTFs");
         Assert.AreEqual(1, cancelled, "Cancelled GLTFs");
     }
-    
-    [UnityTest]
+
+    [UnityTest][Category("ToFix")]
     public IEnumerator TrackFailedGLTFCorrectly()
     {
         gltfLoader = new GLTFSceneImporter("id", null, fileLoader, new GLTFThrottlingCounter(), stream);
         stream.Dispose(); // this should cause an exception
-        
+
         Task task = gltfLoader.LoadScene(cancellationTokenSource.Token);
-        
+
         yield return new WaitUntil(() => task.IsCompleted);
 
         (int loading, int failed, int cancelled, int loaded) = PerformanceAnalytics.GLTFTracker.GetData();
